Refactor: implement task management UI with dialog and grid components
